CORE: - suggestor now disabled by default - better comments - bot doesn't fail due to too many connection but repeats processing for the language instead - main page handling improvements - a bit improved stdout SITE: - improved "what links here" link for disambiguations - other pages text improved for zns page - link to DRDI description in languages comparison table - every language now links the comparison page from its right top corner - fixed table for en. it is possible to process en with limit=1, this is just to be done - other minor fixes
git-svn-id: https://svn.toolserver.org/svnroot/golem@450 2398d10b-674a-4e94-94af-0536ca92ec5b

+++ isolated/handle.sh
@@ -429,9 +429,14 @@ handle ()
            then
              echo server is not allowed for writing\; nothing can be run
            else
-             if [ "$line" != '' ]
+             if [ "${line:0:40}" = 'ERROR 1040 (08004): Too many connections' ]
              then
-               echo -ne $line\\r\\n >> ${language}.debug.log
+               echo "$line" > ${language}.repeat.please
+             else
+               if [ "$line" != '' ]
+               then
+                 echo -ne $line\\r\\n >> ${language}.debug.log
+               fi
              fi
            fi
          fi
@@ -440,7 +445,10 @@ handle ()
        do_stat=0
        echo "$do_stat" > ${language}.no_stat.log
        echo "$do_templates" > ${language}.no_templates.log
-       echo "$line" > stop.please
+       if [ ! -f ${language}.repeat.please ]
+       then
+         echo "$line" > stop.please
+       fi
        ;;
     1) elem=${#fcollection[*]}
        fcollection[$elem]=$line
